+++ src/features/user/components/index.ts
@@ -2,3 +2,5 @@ export { OrderDetailView } from "./OrderDetailView";
 export { UserOrdersView } from "./UserOrdersView";
 export { UserSettingsView } from "./UserSettingsView";
 export { UserAddressesView } from "./UserAddressesView";
+export { UserNotificationsView } from "./UserNotificationsView";
+export { UserEditAddressView } from "./UserEditAddressView";